refactor: migrate retrieval slice into drift-engine with compat stubs
(cherry picked from commit b0887cbb0be9288565330f1c73ac2d2d9e608209)
